+++ lib/datagrid/filters.rb
@@ -5,7 +5,6 @@ module Datagrid
 
     require "datagrid/filters/base_filter"
     require "datagrid/filters/enum_filter"
-    require "datagrid/filters/boolean_enum_filter"
     require "datagrid/filters/extended_boolean_filter"
     require "datagrid/filters/boolean_filter"
     require "datagrid/filters/date_filter"
@@ -22,7 +21,6 @@ module Datagrid
       datetime: Filters::DateTimeFilter,
       string: Filters::StringFilter,
       default: Filters::DefaultFilter,
-      eboolean: Filters::BooleanEnumFilter ,
       xboolean: Filters::ExtendedBooleanFilter ,
       boolean: Filters::BooleanFilter ,
       integer: Filters::IntegerFilter,