+++ lib/datagrid/form_builder.rb
@@ -29,10 +29,6 @@ module Datagrid
     end
 
     protected
-    def datagrid_boolean_enum_filter(attribute_or_filter, options = {})
-      datagrid_enum_filter(attribute_or_filter, options)
-    end
-
     def datagrid_extended_boolean_filter(attribute_or_filter, options = {})
       datagrid_enum_filter(attribute_or_filter, options)
     end