+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master
 
+* Remove deprecated `eboolean` filter
 * Fixed scope wrapping to be universal
 * Deprecated `integer_range_filters` and `date_range_filters`. Use `filter(name, type, range: true)` instead.
 * Add `original_scope` method that returns scope as it was defined without any wrapping [#313](https://github.com/bogdan/datagrid/pull/313)